fixed issues with the buildstat logging stream
Migrating from ora 1 to 3 introduced a couple of small changes within Ora that caused things like string getting written twice because of calling `render` after writing to the stream.  In ora 1 it basically caused the string to be rendered twice overriding the first one.  In ora 3, the string get appended causing strings to be repeated at times.