Adding support for baseUrl in the file watcher system.
The changes here are to start the transition from absolute path to relative paths to a baseUrl. This will make caching more portable as well as all the other file paths in logs, which makes things more stable for writing tests so that they don't fail because absolute paths are very system specific.