Refactored file watching to throttle events so that we are not flooding file bundling.
The approach here is a naive throttle of events. Currently, we support queueing files that need to be bundled when bundling is already in progress. The changes here are an addition to prevent quick file saves from triggering multiple bundle update cycles.  We want to trigger bundle cycles after file events slow down enough.